Move Content Service model classes into their own folder
