+++ src/SourceDocumentCoordinator/SourceDocumentCoordinator/HttpClients/ContentServiceApiClient.cs
@@ -1,15 +1,14 @@
 ﻿using System;
 using System.IO;
-using System.Net;
 using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Text;
 using System.Threading.Tasks;
 using Common.Helpers;
-using Microsoft.EntityFrameworkCore.SqlServer.Query.ExpressionTranslators.Internal;
 using Microsoft.Extensions.Options;
 using Newtonsoft.Json;
 using SourceDocumentCoordinator.Config;
+using SourceDocumentCoordinator.Models;
 
 namespace SourceDocumentCoordinator.HttpClients
 {
