+++ src/SourceDocumentCoordinator/SourceDocumentCoordinator/Models/ContentServiceMetadata.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace SourceDocumentCoordinator.HttpClients
+namespace SourceDocumentCoordinator.Models
 {
     public class ContentServiceMetadata
     {
