+++ src/SourceDocumentCoordinator/SourceDocumentCoordinator/Models/ContentServiceResponse.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace SourceDocumentCoordinator.HttpClients
+﻿namespace SourceDocumentCoordinator.Models
 {
     public class ContentServiceResponse
     {
